### 2019-01-16 #276 @richelbilderbeek
Presubmission inquiry: tmhmm
Submitting Author: Richel J.C. Bilderbeek (@richelbilderbeek)
Repository: https://github.com/richelbilderbeek/tmhmm

---

```
Package: tmhmm
Title: Call TMHMM from R
Version: 1.0
Author: Richel J.C. Bilderbeek
Maintainer: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
  'TMHMM' is a program to predict transmembrane helices in proteins.
  'TMHMM' can be called from a website and/or locally using Bash.
  This package allows to install and call TMHMM from R and parse its results.
License: GPL-3 | file LICENSE
    rappdirs,
    testit
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

The package allows the user to call the tool 'TMHMM' from R,
so it extracts the results from that tool.

Bioinformaticians that use 'TMHMM'. It's used to determine the location (inside,
outside, in the membrane) of amino acids in a membrane protein.

No.

I am unsure about the compatibility between 'TMHMM' and rOpenSci guidelines.

'TMHMM', at [its download page](http://www.cbs.dtu.dk/cgi-bin/nph-sw_request?tmhmm)
is very clear:

> Please note that this page is intended for academic users only.

'TMHMM' ensures that its users are academic users, as they have to fill
in a form [at the download page](http://www.cbs.dtu.dk/cgi-bin/nph-sw_request?tmhmm)
in which they have to specify their academic role. If the user passes this, he/she
obtains a download URL that expires.

To ensure not to hurt 'TMHMM' policy, the users of this package must
specify such an obtained download URL.

So, although the package in its current forms plays nicely with 'TMHMM',
I am unsure if this way is compatible with the rOpenSci guidelines.
I feel it does, but I'll submit the package after confirmation of this.
