### 2019-01-17 #276 @sckott
We just saw https://github.com/dansondergaard/tmhmm.py  which is a separate implementation, and so doesn't require the command line utility. Have you considered wrapping it with [reticulate](https://cran.rstudio.com/web/packages/reticulate/)? I don't know if it will work since the python lib has C code, but seems worth a try. If it works and if it means your pkg works on all platforms, and doesn't require the academic only gate, then that seems like a win, yes?
