### 2019-01-18 #276 @richelbilderbeek
Hi @sckott, thanks for your thoughts!

I fully agree that -in hindsight- I maybe should have used the program by @dansondergaard (yes, I think it's awesome!).

On the other hand, to throw useful (Linux-only) code away just for a potential superior alternative, seems like unpragmatic. Apparently, [ropensci editors are working on accepting bindings to single-OS tools](https://github.com/ropensci/dev_guide/pull/118).

I think binding to a single-OS tool encourages the authors of that tool to broaden their range of OSes.

And back to your point, I think that a package that binds to @dansondergaard's version will be superior and I will happily redirect my users to such a superior package. Or perhaps transition to using it, without hurting my users.

I think the only/biggest showstopper would be the authors of the TMHMM tool I call from R. If they don't want to, I will happily go the extra mile to use @dansondergaards work.
