### 2019-01-23 #276 @noamross
Let me just comment here on how this package fits into overall scope given the proposed new scope definition at https://github.com/ropensci/dev_guide/pull/118

-   Yup, is a scientific and not a general system utility
-   This is a very thin wrapper - calling the program via `system` and returning the results as a string.  There's not any particular data handling/parsing (not that there _needs_ to be)
-   This package partially handles installation.  However, it mostly handles it in terms of putting the install directory in a place easy for the package to handle (In an `rappdirs` directory).  There's not a complex installation process to overcome for this utility - the biggest hurdle, still present when using the R package, is requesting permission to use it.
-   The utility is not open sourced or open licensed, but an alternative open-source version exists.
-   The package only works on a single platform, because the utility is single platform.  The alternative is multi-platform, or this package could attempt to make the utility multi-platform via a Docker/[**stevedore**](https://github.com/richfitz/stevedore) based solution.

Based on the above, I think that on balance, this package doesn't meet our criteria.  It doesn't add significant value over a user handling the system utility themselves, and we can see clear cases - using the alternative utility or another multi-platform, where a different package would supplant this one.

Thoughts @richelbilderbeek or @sckott?
