### 2019-01-23 #276 @richelbilderbeek
Thanks @noamross for the crisp overview over the decision making. I can fully understand would the submission not fulfil the criteria.

But I wonder, regardless of these points, it would save people time (especially beginner R users) to re-invent the wheel. The package does what it is made for well. If the goal of rOpenSci is to help people be more efficient with their time, then a pragmatic acceptance of this package maybe would be the lesser evil. As a use case: I already depend on this package ([the epitopeome package](https://github.com/richelbilderbeek/epitopeome)).

On the other hand, @sckott already gave a superior solution and I feel a rejection would be justified. It would give a clear signal to the TMHMM authors to support more platforms. Sure, I'll have to write a new package, but well, sometimes that's what it is :rainbow:
