### 2019-01-29 #276 @richelbilderbeek
Got this in return:

> All the CBS packages work in the same way, so we would kindly ask you to follow the model we have.
> Hope that can be incorporated in the process of setting it up.

I suggest I retract this submission and use the TMHMM implementation suggested by @sckott.

Agree?
