### 2019-02-06 #276 @dansondergaard
@richelbilderbeek, I believe @sckott referred to the model used for the CBS software, not what the alternative is :)
