### 2019-02-06 #276 @sckott
Agree, makes sense to use https://github.com/dansondergaard/tmhmm.py
