+++ R/get_tmhmm_url.R
@@ -3,5 +3,5 @@
 #' @author Richel J.C. Bilderbeek
 #' @export
 get_tmhmm_url <- function() {
-  "http://www.cbs.dtu.dk/download/D3198788-0F1D-11E9-883C-84B3B9CD16B5/"
+  "http://www.cbs.dtu.dk/download/239C7BA0-5205-11E9-9BC5-D57AB9CD16B5/"
 }
