### 2019-03-29 6b61a25
Improve install script [skip ci]
