### 2019-03-29 a4592d6
Fix script [skip ci]
