### 2019-03-29 1a4cad8
Fix script [skip ci]
