### 2019-03-29 79e92d0
Correct log filename in script [skip ci]
