+++ R/get_tmhmm_url.R
@@ -3,5 +3,5 @@
 #' @author Richel J.C. Bilderbeek
 #' @export
 get_tmhmm_url <- function() {
-  "http://www.cbs.dtu.dk/download/239C7BA0-5205-11E9-9BC5-D57AB9CD16B5/"
+  "http://www.cbs.dtu.dk/download/1E7BD514-5230-11E9-81C3-23DBB9CD16B5/"
 }
