### 2019-04-15 #276 @richelbilderbeek
I suggest to close this Issue, I will create a new one if I submit a package that is FOSS.
